+++ src/components/Description/Description.scss
